Creation of areamaker
areamaker.py is a program that creates areas that can be traversed in ObjectRPG.
This inital version allows creation of an areaset of any dimensions. Each areaset contains a dictionary of areas.
Each area is referenced by the idno field. Each area has a name that is displayed in the game.
Players can traverse the areas in 6 directions: North, South, East, West, Up, and Down.
Currently Up and Down are not fully implemented as they are intended to link to different areasets.
Each area may contain a list of items. This feature is not yet fully implemented.
Each area has an initial message that displays in game when the player first enters the area. The "look" command allows the player to read the initial message again. The initswitch flag keeps track of if the initmessage has been read for the first time or not.
Each area contains a message that is displayed every time the player enters the area.
Each area may also contain a sign which may be read in game using the "read" command.
Each area contains an enemy flag that determines if there will be random battles in that area.
In game, a random chance determines if the battle occurs. this will be modified in the future.
areamaker allows the user to create areas one by one as well as traverse the areaset and edit the current area on the grid. If the area doesn't exist, it is created and opened for editing.
areamaker allows the user to save the areaset by supplying a filename.
Existing areasets may also be loaded.

ObjectRPG has also been updated to be able to load an areaset and traverse it.
A new function called showstatus() displays the area name and various messages pertaining to the area.
New commands for traversing and interacting with the area are implemented. The command "go" will ask for a direction to be supplied, but only the direction is actually required to traverse the areaset. For example, "east" will traverse to the east if there is a path leading east from the current area.
The commands "look" and "read" allow the player to interact with the area by reading the initmessage and the sign message respectively.
Future commands for take items from the area or dropping items from the inventory are planned for the future.
Saving the game will be updated in the future to also save the areaset to maintain a persistent game world.