Added events, updated movement, save and load.
Areamaker
Area directions are now stored as an array with the areaset name and the position so that movement in any direction can load a new areaset if desired.
Area switches and variables are now stored as dictionaries.
Added event creation in areas.
Added event deletion.
Added ability to add actions to events.
Actions for text, item exchange, and gold exchange added.
Interaction with switches and variables and branching choices planned.

ObjectRPG
Movement system updated so that any direction may lead to a new area as desired.
Updated save and load so that the areaset is saved and loaded with the game data.
Minor clean up of some messages.
Clean up of unused code in battle system.
Added events to areas.
Added event handling to process events in the area.
Fixed potion handling so player characters hold potions as items, but the total count is still displayed in battle and status screens.