Update ObjectRPG.py
Major update.

Added new enemies as well as a boss.
Added scripted battles.

Added outline for Game Data concept.
Modified stats of players and enemies.
Added ID numbers to players.

Updated movement between areas to allow the game designer to specify what items resolve a direction status. This replaces only having locked doors.

Added second position to initswitch on areas so that the game can track both if the area has been visited as well as if it is revealed by a map. Map feature update to come later.

Updated events to have ID numbers as well as hidden status.

Changed preconditions to be stored as a dictionary instead of a list.

Added a Battlecommand class.
In a future update battlecommand will be loaded onto classes to define the actions different characters can take in battle.

Added exception handling for loading the initial game file.

Updated class selection so case sensitivity isn't required. This will likely be changed in the future to a choice selection using numbers instead of requiring the player to type out the name of the class.

Updated addition of new characters to add ID numbers and to allow predefined characters to join the party from events.
A future update will allow the level to be set as well.

Fixed a bug with the intro message.

Updated the way directions are displayed so the game designer can specify the message the game uses.

Updated event handling to follow a better order of operations.

Added a function to check the potion count of all players in the party.

Added event preconition handling.

Fixed bugs in save and load functions.

Updated levelup function.

Updated function that adds enemies to battle.

Added battle turn count to battles. This can be used with scripted battles.

Future update will replace attack functions with one universal attack function that handles player and enemy attacks together as well as physical and magic attacks.

Added handling for enemies to have a next form after being defeated.

Added game over function so events can end the game.

Removed debugging comments throughout.